Adds safety-light project
This project is checking the ambient brightness with a photo-
resistor and if it is very dark the motion detection sensor
is testing if there is motion.
If that is the case the LED will be switched on and made brighter
in an iterative process.
If the room is bright enough or there is no more motion, the LED
is switched off again (after a short grace period).