Add power saving mode to safety light
In order to make the battery last longer, the low-power
library has been added. On every iteration we check if the last
time additional light was necessary is longer than 30 seconds ago
and the current brightness is also high enough. In that case,
we enable low power sleep mode for 8 seconds.