Adds implementation of a light seeking car
This smart car has two main behaviours. First the car tries to
come closer do a light source by checking two photo resistors.
In addition an ultrasonic range sensor is used to check if there
is any obstacle in the car's way. If so it tries to find a way
around it.