Fix: servo reset and standup command generation
This change addresses several bugs and minor issues:

Bug 1: server reset was executed on 3 motors only.
---
Due to a missing real implementation we repeatedly reset servos 1-3
instead of all of them.
Fixing the constant definition fixed this bug

Bug 2: Standup sequence generation was broken
---
Due to an issue with the vector range definition, we used a range
of 18 values instead of 6 (motors instead of legs).
Thus we generated invalid command which resulted in wrong leg positions.
This issue was fixed by switching from `initPositions.size()` to
`xPins.size()`. However the remaining issue is, that the code
now assumes that the xPins and zPins are of the same length.
This is a fair assumption, but using something like std::min instead
could make it more solid.

Improvement:
-----
The change adds use of the command group to the Y axis initialization.
This has the advantage, that we get rid of small timing issues, that
we have, when sending multiple commands and not just one single long
one.