feat: cleanup and some documentation
Since the C++11 appraoches work fine for now, most of the
original motor control algorithms are removed.
Effectively switching completely from `move[SingleServo]` to
`moveCommandGroup`.

I left the generation of the x,y,z vectors as a comment, because
I think it might change soon. In addition the better approach
would be to compile the init positions from the vectors
and not the vectors from the initPositions.
In addition this would help us get rid of the std::transform
statement to pull out the first and second elements of the
init positions. (maybe we should invert the nested vector.