feat: Implement first forward walking algo
This change introduces the ability of walking forward.
We use a 2-tripod based approach. The sequence is documented
in the [lynxmotion SSC32-U](http://www.lynxmotion.com/images/data/lynxmotion_ssc-32u_usb_user_guide.pdf)
user guide. The controller implements this sequence directly, but
due to the wiring that would be completely different from the wiring
we did, and in order to learn how to implement it in software,
I decided to implement it on my own.
This version needed a lot of fine tuning regarding the angles and
speed, but it is finally working.
Further adjustments are coming