feat: Measure legs limbs and body
In order to use inverse kinemetics, we need to first know the
length of all the limbs (coxa, femur and tibia), so that we can
calculate the angles for the joints.
This PR encodes all of the measured values as well as the max and min
values of each joint's servo rotation in constant.
One thing we might consider is multiplying all the constants we have
by 10, so that we have 1 decimal without using floats.