feat: Implement inverse kinematics
This commit introduces inverse kinematics (body IK and leg IK).
In order to not broke the existing codes many vector definitions
and constants have been added, instead of adjusting existing once.
The following general changes were necessary:
* We now count legs clockwise, starting with the front-right leg
* The servo default levels and constants were update to match a new
  default angle of 0
* In order to make IK work, we added measurements for all links

This change also adds a testIK method, that is for testing out
if all the angles are generated correctly. The current test of changing
the Z value works as expected and can replace the hard-coded upDown
methods.
However we need to check a perfect right angle is always the most
stable angle. However this is independend of the removal of the
hard coded solution.
Next step: Test out the rotation values rotX, rotY, rotZ, currently
the hexapod lifts the front legs when doing a X-axis rotation, which
is not expected.