Added Servo, DHT-22 and analog signal processing
As part of the AdafruitIO learning tutorial, the following has been
added:

Servo Processing
-----
It is now possible to control a servo angle (0-180) via the adafruitIO
feed.

Analog I/O
----
In order to test analog signal processing an analog input feed has
been added that is based on a photocell that processes the brightess
of the ambient.

For analog output a slider has been added to the dashboard that controls
the brightness of an LED

DHT-22
--------
As another step a DHT-22 has been added to the circuit. There
is a dedicated feed per reading, so that temperature and humidity are
recorded independently. In addition I set up a historam for both
values to have some kind of sensor reading history.