Implement Adabox1 project
Adds an example project for adabox01. The prototype is monitoring
temperature (tmp36 sensor) and indicates via LEDs if the temperature
is below or above a given alarm value. The value can be increased
and decreased via two push buttons.
In addition to a red LED signal a buzzer is also activated when the
alert temperature is exceeded.