Adds NFC/RFID support via MFRC522
Adds means for reading a rfid tag and comparing it to a static
access tag id. The authorized tag-id is defined in config.h and
not part of the commit. Just define
`byte goodTagID[5] = {0x1,0x1,0x1,0x1}` in your config.h with your own
tag id values and everything should work as expected.