Replace 8 cable wiring with one cable/port
In order to reduce the number of cables/digital ports required
to use the keypad, we replaced the wiring with a dedicated analog input
that is connected to a voltage devider.
In order to use the Keypad we replaced the `Keypad` library with the
`OnewireKeypad` library which implements the resistor calculation
required for this approach.